services/repo-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
kafka-python==2.0.2
aiohttp==3.9.3
pydantic==2.6.1
python-dotenv==1.0.1
pytest==8.1.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
httpx==0.27.0
